Extract config code into separate module
This starts to mirror how we organise code in every other repo -
it's confusing for this one to be different.

I've moved the config loading to before init_app like it is in our
other repos. This has no impact because init_app just sets up some
request handlers - but should in theory have access to the config.
